pyproject.toml
[project]
name = "artem-kurs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "fastapi-cache2==0.2.2",
    "fastapi[standart]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "redis>=5.0.0",
    "ruff>=0.12.12",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
